// react/src/__generated__/QuotaSettingModalSetMutation.graphql.ts
export type QuotaScopeInput = {
  hard_limit_bytes?: any | null | undefined;
};
export type QuotaSettingModalSetMutation$variables = {
  props: QuotaScopeInput;
  quota_scope_id: string;
  storage_host_name: string;
};
export type QuotaSettingModalSetMutation$data = {
  readonly set_quota_scope: {
    readonly quota_scope: {
      readonly details: {
        readonly hard_limit_bytes: any | null | undefined;
      };
      readonly id: string;
      readonly quota_scope_id: string;
      readonly storage_host_name: string;
    } | null | undefined;
  } | null | undefined;
};
export type QuotaSettingModalSetMutation = {
  response: QuotaSettingModalSetMutation$data;
  variables: QuotaSettingModalSetMutation$variables;
};